Fix missed in32->int64 in df852120098dc7ba5df4a76126c6297c6d2d1b7b
Tag #709.

+++ security/manager/ssl/nsNSSComponent.cpp
@@ -1540,7 +1540,7 @@ CipherSuiteChangeObserver::Observe(nsISupports* aSubject,
           // are enabled in prefs. They are only used on specific
           // sockets as a part of a fallback mechanism.
           // Only the main thread will change sEnabledWeakCiphers.
-          uint32_t enabledWeakCiphers = sEnabledWeakCiphers;
+          uint64_t enabledWeakCiphers = sEnabledWeakCiphers;
           if (cipherEnabled) {
             enabledWeakCiphers |= ((uint64_t)1 << i);
           } else {
